fix MacOS9 Project (T.Yui)
git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/trunk@539 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ win9x/juliet.cpp
@@ -60,7 +60,7 @@ static	_ROMEO		romeo = {NULL};
 // Ž©‘O‚ÅŒŸõ‚·‚éi–î–ì‚³‚ñ•ûŽ®j
 
 #define PCIBUSDEVFUNC(b, d, f)	(((b) << 8) | ((d) << 3) | (f))
-#define	DEVVEND(v, d)			((v) | ((d) << 16))
+#define	DEVVEND(v, d)			((ULONG)((v) | ((d) << 16)))
 
 static ULONG searchRomeo(void) {
